refactor: API 폴더 main/manage 하위 폴더로 분리 (#29)
사용자 API와 관리자 API를 명확히 구분

- api/main/ : 사용자 대상 API 7개 파일
- api/manage/ : 관리자 API (기존 유지)
- api/axios.js : 공용 인스턴스 (루트 유지)
- manage 폴더 내 파일 2개에 주석 추가
