+++ src/api/manage/getArticleDetail.ts
@@ -1,3 +1,7 @@
+/**
+ * 관리자 게시글 상세 조회
+ * article_id로 특정 게시글의 상세 정보를 조회
+ */
 import instance from '@/api/axios';
 
 export async function getArticleDetail(article_id: number) {
