+++ src/components/main/adaptive/feature/MYP/BookmarkSection.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
 import { useNavigate } from "react-router-dom";
 import BookmarkItem from "@/components/main/adaptive/feature/MYP/BookmarkItem";
-import { deleteSchoolBookmarksAll, deleteSchoolBookmark, fetchSchoolBookmarks } from "@/api/bookmarks";
+import { deleteSchoolBookmarksAll, deleteSchoolBookmark, fetchSchoolBookmarks } from "@/api/main/bookmarks";
 import { useDeviceStore } from "@/stores/deviceStore";
 import useAuthStore from "@/stores/useAuthStore";
