+++ src/components/main/desktop/common/TabBar.jsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 import { NavLink, useNavigate } from "react-router-dom";
 import { useQuery } from "@tanstack/react-query";
 import useAuthStore from "@/stores/useAuthStore";
-import { fetchUnreadCount } from "@/api/notifications";
+import { fetchUnreadCount } from "@/api/main/notifications";
 import bellIcon from "@/assets/icons/notification.svg";
 import userIcon from "@/assets/icons/user.svg";
 import NotificationModal from "@/components/main/adaptive/common/NotificationModal";
