+++ src/pages/main/CBD/CBDPage.jsx
@@ -6,7 +6,7 @@ import MobileHeader from "@/components/main/mobile/common/MobileHeader";
 import MobileTabBar from "@/components/main/mobile/common/MobileTabBar";
 import MobileFooter from "@/components/main/mobile/common/MobileFooter";
 import ClubDetail from "@/components/main/adaptive/feature/CBD/ClubDetail";
-import { fetchClubDetail } from "@/api/articles";
+import { fetchClubDetail } from "@/api/main/articles";
 import { useDeviceStore } from "@/stores/deviceStore";
 
 const CBDPage = () => {
